1Sheeld Fall Detection System
This uses the accelerometer in a smart fone to detect a fall, communicates it to the arduino. The phone also sends out the alerts and necessary actions when a fall is detected (toggle switch, GPS, GSM, email for IOS, and even room for twitter.)

IOS devices don't allow for SMS so the email portion of the code should be turned on.